Add files via upload
This document presents three narratives of a relative complexity pertaining to an imaginary community called Riverforce and its problems. These narratives are classified and illustrate different levels of descriptive complexity of the thinking in terms of the dimension of Structural Complexity, and the property of Structural Variety and Dimensionality.

Recommended citation: Melo, A. T., Renault, L., Ribeiro, R. (2024). Template narrative scenarios for calibrating evaluation framework for Complex Thinking on the dimension of Structural Complexity and the property of Structural Variety and Dimensionality. Coimbra: Centre for Social Studies. This document was produced as part of the project COMPLEX THINKING & AI, Funded by Stefan Pernar