Add files via upload
This document contains development documents (tests, earlier versions of the evaluation framework and chat logs)used and created during the process of developing an evaluation framework for Complex Thinking and its calibration for the Dimension of Structural Complexity and the properties of Structural Variety and Dimensionality

Recommended citation: Melo, A. T., Renault, L., Lopes, P. D., Ribeiro, R. (2024). Development documents and tests for the development and calibration of an evaluation framework for Complex Thinking on the dimension of Structural Complexity and the property of Structural Variety and Dimensionality. Coimbra: Centre for Social Studies. This document was produced as part of the project COMPLEX THINKING & AI, Funded by Stefan Pernar