Add files via upload
This document contains information on the pilot tests with internal and external users for the development of a scaffolding framework for Complex Thinking and its calibration for the Dimension of Structural Complexity and the properties of Structural Variety and Dimensionality and Relationality

Recommended citation: Melo, A. T., Renault, L., Caves, L., Garnett, P., Lopes, P. D., Ribeiro, R., Santos, F. (2024). Development documents and tests for the development and calibration of an evaluation framework for Complex Thinking on the dimension of Structural Complexity and the property of Structural Variety and Dimensionality. Coimbra: Centre for Social Studies. This document was produced as part of the project COMPLEX THINKING & AI, Funded by Stefan Pernar