Add files via upload
This file contains a history of definitions and prompts to support the development of an evaluation framework for the properties of Structural Variety and Dimensionality and Relationality up to version 15 (First development stage). The versions should all be numbered as V1. followed by the number of version on each sheet, e.g. V1.15					

Recommended citation: Melo, A. T. de ., Renault, L., & Duarte Lopes, P. (2025). Archive history of support definitions for the development of an evaluation framework and scaffolding questions for Complex Thinking: Properties of Structural Variety and Dimensionality and Relationality. Coimbra: Centro de Estudos Sociais. This document was produced as part of the project COMPLEX THINKING & AI, Funded by Stefan Pernar