Add files via upload
This document contains the archive of the first development versions of the protocols for scaffolding Complex Thinking.

Recommended citation: Melo, A. T., Renault, L., Lopes, P.D. (2024). Archive of the first development versions of the protocols for scaffolding Complex Thinking. Coimbra: Centre for Social Studies.
This document was produced as part of the project COMPLEX THINKING & AI, Funded by Stefan Pernar